Refactor y autoría
