+++ src/graphdisplay-base/graphdisplay/__init__.py
@@ -1,4 +1,4 @@
 from .graphdisplay import GraphGUI
 from .graph import Graph
-from .tree.auto_balance_tree import AVLTree
-from .tree.binary_search_tree import BinarySearchTree
+from .trees.auto_balance_tree import AVLTree
+from .trees.binary_search_tree import BinarySearchTree
