+++ src/graphdisplay-base/graphdisplay/trees/__init__.py
@@ -1,0 +1,2 @@
+from .binary_search_tree import BinarySearchTree
+from .auto_balance_tree import AVLTree
